move dynamo logging util to separate module
