+++ torchdynamo/convert_frame.py
@@ -11,6 +11,8 @@ from typing import Callable
 import torch
 from torch.fx.graph_module import _forward_from_src as original_forward_from_src
 
+import torchdynamo.logging
+
 from . import config
 from . import exc
 from .allowed_functions import is_allowed
@@ -395,7 +397,7 @@ def _compile(
         output_codes.add(out_code)
 
         log.log(
-            config.INFO_CODE,
+            torchdynamo.logging.INFO_CODE,
             format_bytecode(
                 "ORIGINAL BYTECODE",
                 code.co_name,
@@ -405,7 +407,7 @@ def _compile(
             ),
         )
         log.log(
-            config.INFO_CODE,
+            torchdynamo.logging.INFO_CODE,
             format_bytecode(
                 "MODIFIED BYTECODE",
                 code.co_name,
@@ -423,7 +425,7 @@ def _compile(
         guard_str = "GUARDS:\n"
         guard_str += "\n".join([f" - {str(guard)}" for guard in sorted(output.guards)])
 
-        log.log(config.INFO_CODE, guard_str)
+        log.log(torchdynamo.logging.INFO_CODE, guard_str)
 
         if guard_export_fn is not None:
             guard_export_fn(output.guards)
